+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changes made are:
 - Removed ``dcx.py`` file write feature into output folder function in ``fromdoc.py`` file.
 - Modified ``fromdoc.py`` to include admoninitons directives.
 - Modified ``fromdoc.py`` to include correct detection of code-block to convert it to respective directives.
+- Multiple ``fromdoc.py`` to include the logic for the conversion of multiple ``.docx`` files to multiple ``.rst`` files.
 
 20230122 - v1.8.3
 =================